fixed connection issue

+++ server.js
@@ -1,7 +1,6 @@
 const express = require('express');
 const mysql = require("mysql2");
 const inquirer = require('inquirer');
-const fs = require('fs'); // do i need this?
 
 const PORT = process.env.PORT || 3001;
 const app = express();
@@ -9,29 +8,6 @@ const app = express();
 // Express middleware
 app.use(express.urlencoded({ extended: false }));
 app.use(express.json());
-
-// CALL ONCE SOMEWHERE IN THE BEGINNING OF THE APP
-// const cTable = require('console.table');
-// console.table([
-//   {
-//     name: 'foo',
-//     age: 10
-//   }, {
-//     name: 'bar',
-//     age: 20
-//   }
-// ]);
-
-// GET ONLY TABLE STRING
-// const table = cTable.getTable([
-//     {
-//       name: 'foo',
-//       age: 10
-//     }, {
-//       name: 'bar',
-//       age: 20
-//     }
-//   ]);
 
 // Connect to the database
 const db = mysql.createConnection(
@@ -51,7 +27,6 @@ db.connect(function (err) {
     }
     start()
 })
-
 
 function start() {
     inquirer.prompt([
@@ -149,31 +124,45 @@ async function addDept() {
 
 // delete a department
 async function removeDept() {
-    app.delete('api/department/:id', (req, res) => {
-        const sql = 'DELETE FROM department WHERE id = ?';
-        const params = [req.params.id];
+    inquirer.prompt([
+        {
+            message: 'What is the name of the department?',
+            type: 'input',
+            name: 'dep_name',
+        }
+    ]).then((answerObj) => {
+        if (answerObj.dep_name == '') {
+            res.json({
+                message: 'No department name provided'
+            });
+        } else {
+            app.delete('api/department/:id', (req, res) => {
+                const sql = 'DELETE FROM department WHERE id = ?';
+                const params = [req.params.id];
 
-        db.query(sql, params, (err, result) => {
-            if (err) {
-                res.statusMessage(400).json({ error: res.message });
-            } else if (!result.affectedRows) {
-                res.json({
-                    message: 'Department not found'
+                db.query(sql, params, (err, result) => {
+                    if (err) {
+                        res.statusMessage(400).json({ error: res.message });
+                    } else if (!result.affectedRows) {
+                        res.json({
+                            message: 'Department not found'
+                        });
+                    } else {
+                        res.json({
+                            message: 'deleted',
+                            changes: result.affectedRows,
+                            id: req.params.id
+                        });
+                    }
                 });
-            } else {
-                res.json({
-                    message: 'deleted',
-                    changes: result.affectedRows,
-                    id: req.params.id
-                });
-            }
-        });
-    });
+            });
+        }
+    })
+
     console.table(data)
     console.log('\n')
     start()
 };
-
 
 async function checkRole() {
     app.get('api/role', (req, res) => {
@@ -245,26 +234,40 @@ async function addRole() {
 };
 
 async function removeRole() {
-    app.delete('api/role/:id', (req, res) => {
-        const sql = 'DELETE FROM role WHERE id = ?';
-        const params = [req.params.id];
+    inquirer.prompt([
+        {
+            message: 'What is the title of the role?',
+            type: 'input',
+            name: 'title',
+        }
+    ]).then((answerObj) => {
+        if (answerObj.title == '') {
+            res.json({
+                message: 'No role title provided'
+            });
+        } else {
+            app.delete('api/role/:id', (req, res) => {
+                const sql = 'DELETE FROM role WHERE id = ?';
+                const params = [req.params.id];
 
-        db.query(sql, params, (err, result) => {
-            if (err) {
-                res.statusMessage(400).json({ error: res.message });
-            } else if (!result.affectedRows) {
-                res.json({
-                    message: 'Role not found'
+                db.query(sql, params, (err, result) => {
+                    if (err) {
+                        res.statusMessage(400).json({ error: res.message });
+                    } else if (!result.affectedRows) {
+                        res.json({
+                            message: 'Role not found'
+                        });
+                    } else {
+                        res.json({
+                            message: 'deleted',
+                            changes: result.affectedRows,
+                            id: req.params.id
+                        });
+                    }
                 });
-            } else {
-                res.json({
-                    message: 'deleted',
-                    changes: result.affectedRows,
-                    id: req.params.id
-                });
-            }
-        });
-    });
+            });
+        }
+    })
     console.table(data)
     console.log('\n')
     start()
@@ -343,27 +346,47 @@ async function addEmp() {
     console.log('\n')
     start()
 };
-async function removeEmp() {
-    app.delete('api/employee/:id', (req, res) => {
-        const sql = 'DELETE FROM department WHERE id = ?';
-        const params = [req.params.id];
 
-        db.query(sql, params, (err, result) => {
-            if (err) {
-                res.statusMessage(400).json({ error: res.message });
-            } else if (!result.affectedRows) {
-                res.json({
-                    message: 'Employee not found'
+async function removeEmp() {
+    inquirer.prompt([
+        {
+            message: 'What is the first name of the employee?',
+            type: 'input',
+            name: 'first_name',
+        },
+        {
+            message: 'What is the last_name of this employee?',
+            type: 'input',
+            name: 'last_name',
+        }
+    ]).then((answerObj) => {
+        if (answerObj.first_name == '' && answerObj.last_name == '') {
+            res.json({
+                message: 'No employee name provided'
+            });
+        } else {
+            app.delete('api/employee/:id', (req, res) => {
+                const sql = 'DELETE FROM department WHERE id = ?';
+                const params = [req.params.id];
+
+                db.query(sql, params, (err, result) => {
+                    if (err) {
+                        res.statusMessage(400).json({ error: res.message });
+                    } else if (!result.affectedRows) {
+                        res.json({
+                            message: 'Employee not found'
+                        });
+                    } else {
+                        res.json({
+                            message: 'deleted',
+                            changes: result.affectedRows,
+                            id: req.params.id
+                        });
+                    }
                 });
-            } else {
-                res.json({
-                    message: 'deleted',
-                    changes: result.affectedRows,
-                    id: req.params.id
-                });
-            }
-        });
-    });
+            });
+        }
+    })
     console.table(data)
     console.log('\n')
     start()
@@ -375,7 +398,7 @@ async function removeEmp() {
 
 
 // Close the connection
-connection.end();
+// connection.end();
 
 
 // Default response for any other request (Not Found)
@@ -386,64 +409,3 @@ app.use((req, res) => {
 app.listen(PORT, () => {
     console.log(`Server running on port ${PORT}`);
 });
-// THIS IS THE END OF THE CODE IVE WRITTEN  
-
-// const fs = require('fs');
-// const mysql = require('mysql2');
-// const inquirer = require('inquirer');
-
-// const connection = mysql.createConnection({
-//   host: 'localhost',
-//   user: 'root',
-//   password: 'root'
-// });
-
-// connection.connect(function(err) {
-//   if (err) {
-//     console.error('Error connecting to the database: ' + err.stack);
-//     return;
-//   }
-
-fs.readFile('schema.sql', 'utf-8', function (err, data) {
-    if (err) {
-        console.error('Error reading schema file: ' + err.stack);
-        return;
-    }
-
-    connection.query(data, function (err, results, fields) {
-        if (err) {
-            console.error('Error executing schema file: ' + err.stack);
-            return;
-        }
-
-        console.log('Database and tables created successfully!');
-
-        inquirer.prompt([
-            {
-                type: 'input',
-                name: 'department_name',
-                message: 'Enter the name of a department:'
-            },
-            {
-                type: 'input',
-                name: 'role_title',
-                message: 'Enter the title of a role:'
-            },
-            {
-                type: 'input',
-                name: 'role_salary',
-                message: 'Enter the salary of the role:'
-            },
-            {
-                type: 'input',
-                name: 'employee_first_name',
-                message: 'Enter the first name of an employee:'
-            },
-            {
-                type: 'input',
-                name: 'employee_last_name',
-                message: 'Enter the last name of the employee:'
-            }
-        ]).then(function (answers) {
-            // Insert data into the department table
-            connection.query('INSERT INTO
